email body and header can be enriched by document attributes via placeholders, fix bug with multiple rules apply for one doc, refs #13
