+++ docker-compose.yaml
@@ -3,7 +3,7 @@ services:
     build:
       dockerfile: Dockerfile
       context: .
-    image: carlosz1986/paperless-mailservice:1.8
+    image: carlosz1986/paperless-mailservice:1.6
     volumes:
       - .:/app
       - ./config:/config
