+++ main.go
@@ -1,41 +1,13 @@
 package main
 
 import (
-	"bytes"
-	"crypto/tls"
-	"encoding/base64"
-	"encoding/json"
 	"fmt"
-	"io/ioutil"
 	"log"
 	"math/rand"
-	"mime/quotedprintable"
-	"net/http"
-	"net/smtp"
+	"strconv"
 	"strings"
 	"time"
 )
-
-type Document struct {
-	ID               int    `json:"id"`
-	Title            string `json:"title"`
-	FileName         string `json:"archived_file_name"`
-	OriginalFileName string `json:"original_file_name"`
-	TagIDs           []int  `json:"tags"`
-}
-
-// getFileName returns the archived filename. For encrypted files it uses the original name.
-func (d *Document) getFileName() string {
-	if d.FileName != "" {
-		return d.FileName
-	}
-	return d.OriginalFileName
-}
-
-type Tag struct {
-	ID   int    `json:"id"`
-	Name string `json:"name"`
-}
 
 func main() {
 	LoadConfig()
@@ -69,6 +41,26 @@ func processJob() error {
 		return fmt.Errorf("error getting tags: %v", err)
 	}
 
+	correspondents, err := getCorrespondents()
+	if err != nil {
+		return fmt.Errorf("error getting correspondents: %v", err)
+	}
+
+	documentTypes, err := getDocumentTypes()
+	if err != nil {
+		return fmt.Errorf("error getting document types: %v", err)
+	}
+
+	storagePaths, err := getStoragePaths()
+	if err != nil {
+		return fmt.Errorf("error getting storage pathes: %v", err)
+	}
+
+	users, err := getUsers()
+	if err != nil {
+		return fmt.Errorf("error getting users: %v", err)
+	}
+
 	processedTag := getTagByName(tags, Config.Paperless.ProcessedTagName)
 	if processedTag == nil {
 		return fmt.Errorf("error finding processedTagName:%s in list from server", Config.Paperless.ProcessedTagName)
@@ -93,6 +85,8 @@ func processJob() error {
 	for _, doc := range documents {
 		// check if rule for document exists, and process doc
 		// all tags of a rule need to be available for the doc
+		atLeastOneRuleMatches := false
+
 		for _, rule := range Config.Paperless.Rules {
 			docMatchesRule := false
 			for _, ruleTag := range rule.Tags {
@@ -121,22 +115,90 @@ func processJob() error {
 			// found a rule that matches, start processing
 			log.Printf("found Rule: %s, that matches Tag(s) (%s) in document: '%s' (%d)", rule.Name, strings.Join(rule.Tags, ","), doc.getFileName(), doc.ID)
 
-			if err := SendProcessDoc(doc, processedTag, rule.ReceiverAddress); err != nil {
+			user := getUserByID(users, doc.OwnerId)
+			if user == nil {
+				log.Printf("warning: could not find user for doc with id=%d, placeholders won't work", doc.ID)
+			}
+
+			correspondent := getCorrespondentByID(correspondents, doc.CorrespondentId)
+			if correspondent == nil {
+				log.Printf("warning: could not find a correspondent for doc with id=%d, placeholders won't work", doc.ID)
+			}
+
+			documentType := getDocumentTypeByID(documentTypes, doc.DocumentTypeId)
+			if documentType == nil {
+				log.Printf("warning: could not find a document type for doc with id=%d, placeholders won't work", doc.ID)
+			}
+
+			storagePath := getStoragePathByID(storagePaths, doc.StoragePath)
+			if storagePath == nil {
+				log.Printf("warning: could not find a storage path for doc with id=%d, placeholders won't work", doc.ID)
+			}
+
+			mailHeader := prepareMail(Config.Email.MailHeader, rule.MailHeader, user, correspondent, documentType, storagePath, &doc)
+			mailBody := prepareMail(Config.Email.MailBody, rule.MailBody, user, correspondent, documentType, storagePath, &doc)
+
+			if err := SendProcessDoc(doc, processedTag, mailHeader, mailBody, rule.ReceiverAddress); err != nil {
 				log.Printf("error processing Doc: %v", err)
 				continue
 			}
 
 			log.Printf("document '%s' (%d) succesfully sent to '%s'", doc.getFileName(), doc.ID, rule.ReceiverAddress)
-			goto nextDoc
+			atLeastOneRuleMatches = true
+			//goto nextDoc
 		}
-		log.Printf("document '%s' (%d) marked for processing, but no Ruleset matches the tags ...", doc.getFileName(), doc.ID)
-	nextDoc:
+		if !atLeastOneRuleMatches {
+			log.Printf("document '%s' (%d) marked for processing, but no Ruleset matches the tags ...", doc.getFileName(), doc.ID)
+		}
+		//nextDoc:
 	}
 
 	return nil
 }
 
-func SendProcessDoc(doc Document, processedTag *Tag, receiverAddress string) error {
+func prepareMail(str, ruleStr string, user *User, correspondent *Correspondent, documenType *DocumentType, storagePath *StoragePath, document *Document) string {
+	// use the header,body string from rule if set
+	if ruleStr != "" {
+		str = ruleStr
+	}
+
+	if user != nil {
+		str = strings.ReplaceAll(str, "%user_name%", user.Username)
+		str = strings.ReplaceAll(str, "%user_id%", strconv.Itoa(user.ID))
+		str = strings.ReplaceAll(str, "%user_email%", user.Email)
+		str = strings.ReplaceAll(str, "%first_name%", user.FirstName)
+		str = strings.ReplaceAll(str, "%last_name%", user.LastName)
+	}
+
+	if correspondent != nil {
+		str = strings.ReplaceAll(str, "%correspondent_name%", correspondent.Name)
+		str = strings.ReplaceAll(str, "%correspondent_id%", strconv.Itoa(correspondent.ID))
+
+	}
+
+	if documenType != nil {
+		str = strings.ReplaceAll(str, "%document_type_name%", documenType.Name)
+		str = strings.ReplaceAll(str, "%document_type_id%", strconv.Itoa(documenType.ID))
+
+	}
+
+	if storagePath != nil {
+		str = strings.ReplaceAll(str, "%storage_path_name%", storagePath.Name)
+		str = strings.ReplaceAll(str, "%storage_path_id%", strconv.Itoa(storagePath.ID))
+		str = strings.ReplaceAll(str, "%storage_path%", storagePath.Path)
+	}
+
+	str = strings.ReplaceAll(str, "%document_id%", strconv.Itoa(document.ID))
+	str = strings.ReplaceAll(str, "%document_url%", document.getDocumentURL())
+	str = strings.ReplaceAll(str, "%document_title%", document.Title)
+	str = strings.ReplaceAll(str, "%document_file_name%", document.getFileName())
+	str = strings.ReplaceAll(str, "%document_created_at%", document.CreatedAt)
+	str = strings.ReplaceAll(str, "%document_modified_at%", document.ModifiedAt)
+
+	return str
+}
+
+func SendProcessDoc(doc Document, processedTag *Tag, mailHeader, mailBody, receiverAddress string) error {
 	// download document
 	bytes, err := downloadDocumentBinary(doc)
 	if err != nil {
@@ -153,8 +215,8 @@ func SendProcessDoc(doc Document, processedTag *Tag, receiverAddress string) err
 		Config.Email.SMTPUser,
 		Config.Email.SMTPPassword,
 		receiverAddress,
-		Config.Email.MailHeader,
-		Config.Email.MailBody,
+		mailHeader,
+		mailBody,
 		doc.getFileName(),
 		bytes)
 
@@ -165,356 +227,6 @@ func SendProcessDoc(doc Document, processedTag *Tag, receiverAddress string) err
 	err = addTagToDocument(doc, *processedTag)
 	if err != nil {
 		return fmt.Errorf("could not add Tag for document '%s' (%d): %v", doc.getFileName(), doc.ID, err)
-	}
-	return nil
-}
-
-func getTags() ([]Tag, error) {
-	var tags []Tag
-	page := 1
-
-	for {
-		url := fmt.Sprintf("%stags/?page=%d", Config.Paperless.InstanceURL, page)
-		req, err := http.NewRequest("GET", url, nil)
-		if err != nil {
-			return nil, err
-		}
-
-		req.Header.Set("Authorization", fmt.Sprintf("Token %s", Config.Paperless.InstanceToken))
-
-		client := &http.Client{}
-		resp, err := client.Do(req)
-		if err != nil {
-			return nil, err
-		}
-		defer resp.Body.Close()
-
-		if resp.StatusCode != http.StatusOK {
-			return nil, fmt.Errorf("failed to fetch tags: %s", resp.Status)
-		}
-
-		var result struct {
-			Results []Tag  `json:"results"`
-			Next    string `json:"next"`
-		}
-
-		err = json.NewDecoder(resp.Body).Decode(&result)
-		if err != nil {
-			return nil, err
-		}
-
-		for _, tag := range result.Results {
-			tags = append(tags, tag)
-		}
-
-		if result.Next == "" {
-			break
-		}
-		page++
-	}
-
-	return tags, nil
-}
-
-func getDocumentsByTag(tag Tag, processedTag Tag) ([]Document, error) {
-	var documents []Document
-	page := 1
-
-	for {
-		url := fmt.Sprintf("%sdocuments/?page=%d&tags__id__all=%d&tags__id__none=%d", Config.Paperless.InstanceURL, page, tag.ID, processedTag.ID)
-		req, err := http.NewRequest("GET", url, nil)
-		if err != nil {
-			return nil, err
-		}
-
-		req.Header.Set("Authorization", fmt.Sprintf("Token %s", Config.Paperless.InstanceToken))
-
-		client := &http.Client{}
-		resp, err := client.Do(req)
-		if err != nil {
-			return nil, err
-		}
-		defer resp.Body.Close()
-
-		if resp.StatusCode != http.StatusOK {
-			return nil, fmt.Errorf("failed to fetch documents: %s", resp.Status)
-		}
-
-		var result struct {
-			Results []Document `json:"results"`
-			Next    string     `json:"next"`
-		}
-
-		err = json.NewDecoder(resp.Body).Decode(&result)
-		if err != nil {
-			return nil, err
-		}
-
-		documents = append(documents, result.Results...)
-
-		if result.Next == "" {
-			break
-		}
-		page++
-	}
-
-	return documents, nil
-}
-
-func addTagToDocument(document Document, tag Tag) error {
-	url := fmt.Sprintf("%sdocuments/bulk_edit/", Config.Paperless.InstanceURL)
-
-	type payload struct {
-		Documents  []int          `json:"documents"`
-		Method     string         `json:"method"`
-		Parameters map[string]int `json:"parameters"`
-	}
-	p, b := payload{
-		Documents:  []int{document.ID},
-		Method:     "add_tag",
-		Parameters: map[string]int{"tag": tag.ID},
-	}, new(bytes.Buffer)
-
-	err := json.NewEncoder(b).Encode(p)
-	if err != nil {
-		return fmt.Errorf("failed to marshal payload: %w", err)
-	}
-
-	// Create the request
-	req, err := http.NewRequest("POST", url, b)
-	if err != nil {
-		return fmt.Errorf("failed to create request: %w", err)
-	}
-
-	// Set the necessary headers
-	req.Header.Set("Content-Type", "application/json")
-	req.Header.Set("Authorization", fmt.Sprintf("Token %s", Config.Paperless.InstanceToken))
-
-	// Send the request
-	client := &http.Client{}
-	resp, err := client.Do(req)
-	if err != nil {
-		return fmt.Errorf("failed to send request: %w", err)
-	}
-	defer resp.Body.Close()
-
-	// Check the response status
-	if resp.StatusCode != http.StatusOK {
-		return fmt.Errorf("bulk Edit failed, Unexpected server status code: %d", resp.StatusCode)
-	}
-
-	return nil
-}
-
-func downloadDocumentBinary(doc Document) ([]byte, error) {
-	url := fmt.Sprintf("%sdocuments/%d/download/", Config.Paperless.InstanceURL, doc.ID)
-	req, err := http.NewRequest("GET", url, nil)
-	if err != nil {
-		return nil, err
-	}
-
-	req.Header.Set("Authorization", fmt.Sprintf("Token %s", Config.Paperless.InstanceToken))
-
-	client := &http.Client{}
-	resp, err := client.Do(req)
-	if err != nil {
-		return nil, err
-	}
-	defer resp.Body.Close()
-
-	if resp.StatusCode != http.StatusOK {
-		return nil, fmt.Errorf("failed to download document ID %d: %s", doc.ID, resp.Status)
-	}
-
-	body, err := ioutil.ReadAll(resp.Body)
-	if err != nil {
-		return nil, err
-	}
-	return body, nil
-}
-
-func getTagByID(tags []Tag, id int) *Tag {
-	for _, tag := range tags {
-		if tag.ID == id {
-			return &tag
-		}
-	}
-	return nil
-}
-
-func getTagByName(tags []Tag, name string) *Tag {
-	for _, tag := range tags {
-		if tag.Name == name {
-			return &tag
-		}
-	}
-	return nil
-}
-
-func toQuotedPrintable(s string) (string, error) {
-	var ac bytes.Buffer
-	w := quotedprintable.NewWriter(&ac)
-
-	if _, err := w.Write([]byte(s)); err != nil {
-		return "", err
-	}
-	if err := w.Close(); err != nil {
-		return "", err
-	}
-
-	return ac.String(), nil
-}
-
-func SendEmailWithPDFBinaryAttachment(smtpHost, smtpPort, connectionType, sender, user, password, recipient, subject, body, filename string, attachment []byte) error {
-	// subject,body to quoted printable
-	subjectP, err := toQuotedPrintable(subject)
-	if err != nil {
-		return err
-	}
-
-	bodyP, err := toQuotedPrintable(body)
-	if err != nil {
-		return err
-	}
-
-	// Create the email header
-	header := make(map[string]string)
-	header["From"] = sender
-	header["To"] = recipient
-	header["Subject"] = fmt.Sprintf("=?UTF-8?q?%s?=", subjectP)
-	header["MIME-Version"] = "1.0"
-	header["Content-Type"] = `multipart/mixed; boundary="BOUNDARY"`
-	header["Date"] = time.Now().Format(time.RFC1123Z)
-
-	var emailBuf bytes.Buffer
-	for k, v := range header {
-		emailBuf.WriteString(fmt.Sprintf("%s: %s\r\n", k, v))
-	}
-	emailBuf.WriteString("\r\n--BOUNDARY\r\n")
-
-	// Create the body part
-	emailBuf.WriteString(fmt.Sprintf(`Content-Type: text/plain; charset="UTF-8"; boundary="BOUNDARY"%s`, "\r\n"))
-	emailBuf.WriteString("Content-Transfer-Encoding: quoted-printable\r\n\r\n")
-
-	// write the email body to the buffer
-	emailBuf.WriteString(bodyP)
-	emailBuf.WriteString("\r\n--BOUNDARY\r\n")
-
-	// Create the attachment part
-	emailBuf.WriteString(fmt.Sprintf(`Content-Type: application/pdf; name="%s"%s`, filename, "\r\n"))
-	emailBuf.WriteString("Content-Transfer-Encoding: base64\r\n")
-	emailBuf.WriteString(fmt.Sprintf(`Content-Disposition: attachment; filename="%s"%s`, filename, "\r\n\r\n"))
-
-	b := make([]byte, base64.StdEncoding.EncodedLen(len(attachment)))
-	base64.StdEncoding.Encode(b, attachment)
-
-	if err := addLinesSplittedToBuffer(b, &emailBuf); err != nil {
-		return fmt.Errorf("failed to add line separators to BinaryFile: %v", err)
-	}
-
-	emailBuf.WriteString("\r\n--BOUNDARY--")
-
-	addr := fmt.Sprintf("%s:%s", smtpHost, smtpPort)
-	auth := smtp.PlainAuth("", user, password, smtpHost)
-
-	var client *smtp.Client
-
-	if connectionType == "tls" {
-
-		// Create an SSL/TLS connection
-		tlsConfig := &tls.Config{
-			InsecureSkipVerify: false,
-			ServerName:         smtpHost,
-		}
-
-		conn, err := tls.Dial("tcp", addr, tlsConfig)
-		if err != nil {
-			if err.Error() == "tls: first record does not look like a TLS handshake" {
-				return fmt.Errorf("failed to dial TLS: %v - Try to change smtpConnectionType Config", err)
-			}
-			return fmt.Errorf("failed to dial TLS: %v", err)
-		}
-
-		// Create new client using the SSL connection
-		client, err = smtp.NewClient(conn, smtpHost)
-		if err != nil {
-			return fmt.Errorf("failed to create SMTP client: %v", err)
-		}
-		defer client.Close()
-
-		// Authenticate
-		if err = client.Auth(auth); err != nil {
-			return fmt.Errorf("failed to authenticate: %v", err)
-		}
-
-	} else if connectionType == "starttls" {
-		// Handle TLS/STARTTLS (port 587)
-		client, err = smtp.Dial(addr)
-		if err != nil {
-			if err.Error() == "EOF" {
-				return fmt.Errorf("failed to dial: %v - Try to change smtpConnectionType Config", err)
-			}
-			return fmt.Errorf("failed to dial: %v", err)
-		}
-		defer client.Close()
-
-		// TLS
-		tlsConfig := &tls.Config{
-			InsecureSkipVerify: false,
-			ServerName:         smtpHost,
-		}
-		if err = client.StartTLS(tlsConfig); err != nil {
-			return fmt.Errorf("failed to start TLS: %v", err)
-		}
-
-		// Authenticate
-		if err = client.Auth(auth); err != nil {
-			return fmt.Errorf("failed toauthenticate: %v", err)
-		}
-	} else {
-		return fmt.Errorf("given SMTP connection Type invalid")
-	}
-
-	// Set the sender and recipient
-	if err := client.Mail(sender); err != nil {
-		return fmt.Errorf("failed to set mail sender: %v", err)
-	}
-
-	if err := client.Rcpt(recipient); err != nil {
-		return fmt.Errorf("failed to set mail receiver: %v", err)
-	}
-
-	// Get the data writer to send the email body
-	w, err := client.Data()
-	if err != nil {
-		return fmt.Errorf("failed to get writer: %v", err)
-	}
-	defer w.Close()
-
-	//log.Print(emailBuf.String())
-	// Write the message to the data writer
-	_, err = w.Write(emailBuf.Bytes())
-	if err != nil {
-		return fmt.Errorf("unable to send email: %v", err)
-	}
-
-	client.Quit()
-
-	return nil
-}
-
-func addLinesSplittedToBuffer(b []byte, emailBuf *bytes.Buffer) error {
-	// limit the lines to up to 76 chars
-	for i, l := 0, len(b); i < l; i++ {
-		if err := emailBuf.WriteByte(b[i]); err != nil {
-			return err
-		}
-		if (i+1)%76 == 0 {
-			if _, err := emailBuf.WriteString("\r\n"); err != nil {
-				return err
-			}
-
-		}
 	}
 	return nil
 }
